requirements.txt
tensorflow
pillow
opencv-python
scikit-learn
numpy
matplotlib
flask
dill
catboost
xgboost
-e .